pyproject.toml
[project]
name = "fastpdfreport"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "celery>=5.5.3",
    "fastapi[standard]>=0.115.13",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "pydantic-settings>=2.10.0",
    "pyjwt>=2.10.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "uvloop>=0.21.0; sys_platform != 'win32'", # для скорости работы fastapi
]

[tool.ruff]
ignore = ["F403", "F405"]
